Added Nodes
Added Nodes that count up the number of resources to be distributed.

Currently two nodes begin as if owned by Player 1, and thus is awarded resources based on adjacent tiles, which can be viewed as a log in the Console.

Pressing the 'Update Board' button will reset the nodes to blank, and therefore no resources will be awarded.

Next to work on:
1) determining if the maximum node count has been reached, and if so, do not distribute resources.
2) Got the 'dots' that indicate maximum node count to show up on tiles.
3) Add to score and make placeholder text to show resources and points.
4) Add branches.
5) Add ability to click on an empty node to place a new node.