Added Node Functionality
- Added 'dots' to each tile to indicate the maximum number of nodes allowed on said tile
- Added functionality to the nodes as described below
- Added text objects which will show the amount of resources collected

Clicking on a node will now turn the node Orange, and if you click the node again before clicking the 'Make Move' button, it will turn gray again. However, if you click the 'Make Move' button while having one or more new nodes clicked to orange, they will be solidified as Player1's node and will begin contibuting to the player's resources every time the 'Make Move' button is pressed.

If the amount of Nodes on a tile exceeds the limit shown on the tile itself, no resources will be produced -- I plan on making some changes to the sprite to indicate that a tile is blocked.

To do:
1) Add cost to making nodes
2) Add branches
3) Figure out how the heck to capture tiles with branches...